requirements-osx.txt
pyfrc
wpilib.kinematics
robotpy-navx
robotpy-commands-v1
robotpy-rev-color
# robotpy-ctre
# robotpy-rev